refactor: Move `ReservationKeylessEntry` to its own file
