+++ apps/admin-ui/src/spa/reservations/[id]/index.tsx
@@ -3,7 +3,6 @@ import { trim } from "lodash";
 import React, { useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useParams, useSearchParams } from "react-router-dom";
-import styled from "styled-components";
 import type { TFunction } from "i18next";
 import { add, startOfISOWeek } from "date-fns";
 import {
@@ -11,8 +10,6 @@ import {
   CustomerTypeChoice,
   type ReservationQuery,
   ReservationStateChoice,
-  useChangeReservationAccessCodeMutation,
-  useRepairReservationAccessCodeMutation,
   useReservationQuery,
   UserPermissionChoice,
 } from "@gql/gql-types";
@@ -38,65 +35,13 @@ import { useRecurringReservations } from "@/hooks";
 import { ApprovalButtonsRecurring } from "./ApprovalButtonsRecurring";
 import ReservationTitleSection from "./ReservationTitleSection";
 import { base64encode, isPriceFree } from "common/src/helpers";
-import { formatAgeGroup, formatTime } from "@/common/util";
+import { formatAgeGroup } from "@/common/util";
 import Error404 from "@/common/Error404";
-import {
-  Accordion as AccordionBase,
-  Button,
-  ButtonSize,
-  IconAlertCircleFill,
-} from "hds-react";
-import {
-  ApplicationDatas,
-  KVWrapper,
-  Label,
-  Summary,
-  SummaryFourColumns,
-  Value,
-} from "@/styles/util";
-import { errorToast, successToast } from "common/src/common/toast";
-import { getValidationErrors } from "common/src/apolloUtils";
+import { ApplicationDatas, Summary } from "@/styles/util";
+import { Accordion, DataWrapper } from "./components";
+import { ReservationKeylessEntry } from "./ReservationKeylessEntrySection";
 
 type ReservationType = NonNullable<ReservationQuery["reservation"]>;
-
-const Accordion = styled(AccordionBase).attrs({
-  closeButton: false,
-})`
-  > div > div:not([class^="LoadingSpinner-module_loadingSpinner"]) {
-    width: 100%;
-  }
-
-  && {
-    --icon-size: 24px;
-
-    [class^="Accordion-module_accordionHeader__"] {
-      --icon-size: 32px;
-    }
-  }
-`;
-
-function DataWrapper({
-  label,
-  children,
-  isWide,
-  isSummary,
-}: Readonly<{
-  label: string;
-  children: React.ReactNode;
-  isWide?: boolean;
-  isSummary?: boolean;
-}>): JSX.Element {
-  const testSection = isSummary ? "summary" : "info";
-  const testId = `reservation__${testSection}--${label}`;
-  return (
-    <KVWrapper $isWide={isWide} $isSummary={isSummary}>
-      <Label $isSummary={isSummary}>{label}:</Label>
-      <Value data-testid={testId} $isSummary={isSummary}>
-        {children}
-      </Value>
-    </KVWrapper>
-  );
-}
 
 function ButtonsWithPermChecks({
   reservation,
@@ -263,103 +208,6 @@ function ReservationSummary({
           </DataWrapper>
         ))}
     </Summary>
-  );
-}
-
-export function ReservationKeylessEntry({
-  reservation,
-  onSuccess,
-}: Readonly<{
-  reservation: ReservationType;
-  onSuccess: () => void;
-}>) {
-  const { t, i18n } = useTranslation();
-  const [changeAccessCodeMutation] = useChangeReservationAccessCodeMutation();
-  const [repairAccessCodeAccessCodeMutation] =
-    useRepairReservationAccessCodeMutation();
-
-  const handleButton = async (reservationPk: number) => {
-    const payload = { variables: { input: { pk: reservationPk } } };
-
-    try {
-      if (
-        reservation.pindoraInfo?.accessCodeIsActive ===
-        reservation.accessCodeShouldBeActive
-      ) {
-        await changeAccessCodeMutation(payload);
-        successToast({
-          text: t("RequestedReservation.accessCodeChangedSuccess"),
-        });
-      } else {
-        await repairAccessCodeAccessCodeMutation(payload);
-        successToast({
-          text: t("RequestedReservation.accessCodeRepairedSuccess"),
-        });
-      }
-      onSuccess(); // refetch reservation
-    } catch (err: unknown) {
-      handleError(err);
-    }
-  };
-
-  const handleError = (e: unknown) => {
-    const validationErrors = getValidationErrors(e);
-    if (validationErrors.length > 0) {
-      const code = validationErrors[0].validation_code;
-      if (code && i18n.exists(`errors.backendValidation.${code}`)) {
-        errorToast({ text: t(`errors.backendValidation.${code}`) });
-        return;
-      }
-      errorToast({
-        text: validationErrors[0].message ?? validationErrors[0].code,
-      });
-      return;
-    }
-
-    if (e instanceof Error) {
-      errorToast({ text: e.message });
-    } else {
-      errorToast({ text: t("errors.descriptive.genericError") });
-    }
-  };
-
-  return (
-    <Accordion
-      id="reservation__access-type"
-      heading={t("RequestedReservation.keylessEntry")}
-      initiallyOpen={false}
-    >
-      <div>
-        <SummaryFourColumns>
-          <DataWrapper label={t("RequestedReservation.accessCodeLabel")}>
-            {reservation.pindoraInfo?.accessCode ?? "-"}
-          </DataWrapper>
-          <DataWrapper label={t("RequestedReservation.accessCodeStatusLabel")}>
-            {reservation.pindoraInfo?.accessCodeIsActive
-              ? t("RequestedReservation.accessCodeStatusActive")
-              : t("RequestedReservation.accessCodeStatusInactive")}
-            {reservation.pindoraInfo?.accessCodeIsActive !==
-              reservation.accessCodeShouldBeActive && <IconAlertCircleFill />}
-          </DataWrapper>
-          <DataWrapper
-            label={t("RequestedReservation.accessCodeValidityLabel")}
-          >
-            {reservation.pindoraInfo
-              ? `${formatTime(reservation.pindoraInfo.accessCodeBeginsAt)}–${formatTime(reservation.pindoraInfo.accessCodeEndsAt)}`
-              : "-"}
-          </DataWrapper>
-          <Button
-            size={ButtonSize.Small}
-            onClick={() => handleButton(reservation.pk ?? 0)}
-          >
-            {reservation.pindoraInfo?.accessCodeIsActive ===
-            reservation.accessCodeShouldBeActive
-              ? t("RequestedReservation.accessCodeChange")
-              : t("RequestedReservation.accessCodeRepair")}
-          </Button>
-        </SummaryFourColumns>
-      </div>
-    </Accordion>
   );
 }
